Pulls out ImageTextField
