+++ 12AM/Base.lproj/Main.storyboard
@@ -114,7 +114,7 @@
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Username" textAlignment="natural" minimumFontSize="17" background="EmailLine" translatesAutoresizingMaskIntoConstraints="NO" id="Atf-5l-PGl" customClass="DesignableTextField" customModule="_2AM" customModuleProvider="target">
+                            <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Username" textAlignment="natural" minimumFontSize="17" background="EmailLine" translatesAutoresizingMaskIntoConstraints="NO" id="Atf-5l-PGl" customClass="ImageTextField" customModule="_2AM" customModuleProvider="target">
                                 <rect key="frame" x="16" y="273" width="343" height="30"/>
                                 <color key="tintColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                                 <constraints>
@@ -153,7 +153,7 @@
                                     <constraint firstAttribute="width" constant="115" id="uZO-uY-Jjs"/>
                                 </constraints>
                             </imageView>
-                            <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="email@me.com" textAlignment="natural" minimumFontSize="17" background="EmailLine" translatesAutoresizingMaskIntoConstraints="NO" id="Q43-6v-M7g" customClass="DesignableTextField" customModule="_2AM" customModuleProvider="target">
+                            <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="email@me.com" textAlignment="natural" minimumFontSize="17" background="EmailLine" translatesAutoresizingMaskIntoConstraints="NO" id="Q43-6v-M7g" customClass="ImageTextField" customModule="_2AM" customModuleProvider="target">
                                 <rect key="frame" x="16" y="327" width="343" height="30"/>
                                 <color key="backgroundColor" red="0.93533932743766757" green="0.97384597081218272" blue="0.96503649470140263" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                 <color key="tintColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
